// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type AcfLink = {
  __typename?: 'AcfLink';
  /** The target of the link (_blank, etc) */
  target?: Maybe<Scalars['String']['output']>;
  /** The title of the link */
  title?: Maybe<Scalars['String']['output']>;
  /** The url of the link */
  url?: Maybe<Scalars['String']['output']>;
};

/** Connection between the Config_Fields type and the MediaItem type */
export type AcfMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'AcfMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type Config = AcfFieldGroup & AcfFieldGroupFields & Config_Fields & {
  __typename?: 'Config';
  /** Field of the &quot;image&quot; Field Type added to the schema as part of the &quot;Config&quot; Field Group */
  brandLogo?: Maybe<AcfMediaItemConnectionEdge>;
  /** Field of the &quot;url&quot; Field Type added to the schema as part of the &quot;Config&quot; Field Group */
  ctaLink?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;Config&quot; Field Group */
  ctaText?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type DatabaseIdentifier = {
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
};
// ...
export type GeneralSettings = {
  __typename?: 'GeneralSettings';
  /** A date format for all date strings. */
  dateFormat?: Maybe<Scalars['String']['output']>;
  /** Site tagline. */
  description?: Maybe<Scalars['String']['output']>;
  /** This address is used for admin purposes, like new user notification. */
  email?: Maybe<Scalars['String']['output']>;
  /** WordPress locale code. */
  language?: Maybe<Scalars['String']['output']>;
  /** A day number of the week that the week should start on. */
  startOfWeek?: Maybe<Scalars['Int']['output']>;
  /** A time format for all time strings. */
  timeFormat?: Maybe<Scalars['String']['output']>;
  /** A city in the same timezone as you. */
  timezone?: Maybe<Scalars['String']['output']>;
  /** Site title. */
  title?: Maybe<Scalars['String']['output']>;
  /** Site URL. */
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type Menu = DatabaseIdentifier & Node & {
  __typename?: 'Menu';
  /** The number of items in the menu */
  count?: Maybe<Scalars['Int']['output']>;
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
  /** The globally unique identifier of the nav menu object. */
  id: Scalars['ID']['output'];
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']['output']>;
  /** The locations a menu is assigned to */
  locations?: Maybe<Array<Maybe<MenuLocationEnum>>>;
  /**
   * WP ID of the nav menu.
   * @deprecated Deprecated in favor of the databaseId field
   */
  menuId?: Maybe<Scalars['Int']['output']>;
  /** Connection between the Menu type and the MenuItem type */
  menuItems?: Maybe<MenuToMenuItemConnection>;
  /** Display name of the menu. Equivalent to WP_Term-&gt;name. */
  name?: Maybe<Scalars['String']['output']>;
  /** The url friendly name of the menu. Equivalent to WP_Term-&gt;slug */
  slug?: Maybe<Scalars['String']['output']>;
};
// ...
export type MenuItem = DatabaseIdentifier & Node & {
  __typename?: 'MenuItem';
  /** Connection between the MenuItem type and the MenuItem type */
  childItems?: Maybe<MenuItemToMenuItemConnection>;
  /** Connection from MenuItem to it&#039;s connected node */
  connectedNode?: Maybe<MenuItemToMenuItemLinkableConnectionEdge>;
  /**
   * The object connected to this menu item.
   * @deprecated Deprecated in favor of the connectedNode field
   */
  connectedObject?: Maybe<MenuItemObjectUnion>;
  /** Class attribute for the menu item link */
  cssClasses?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int']['output'];
  /** Description of the menu item. */
  description?: Maybe<Scalars['String']['output']>;
  /** The globally unique identifier of the nav menu item object. */
  id: Scalars['ID']['output'];
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']['output']>;
  /** Label or title of the menu item. */
  label?: Maybe<Scalars['String']['output']>;
  /** Link relationship (XFN) of the menu item. */
  linkRelationship?: Maybe<Scalars['String']['output']>;
  /** The locations the menu item&#039;s Menu is assigned to */
  locations?: Maybe<Array<Maybe<MenuLocationEnum>>>;
  /** The Menu a MenuItem is part of */
  menu?: Maybe<MenuItemToMenuConnectionEdge>;
  /**
   * WP ID of the menu item.
   * @deprecated Deprecated in favor of the databaseId field
   */
  menuItemId?: Maybe<Scalars['Int']['output']>;
  /** Menu item order */
  order?: Maybe<Scalars['Int']['output']>;
  /** The database id of the parent menu item or null if it is the root */
  parentDatabaseId?: Maybe<Scalars['Int']['output']>;
  /** The globally unique identifier of the parent nav menu item object. */
  parentId?: Maybe<Scalars['ID']['output']>;
  /** Path for the resource. Relative path for internal resources. Absolute path for external resources. */
  path?: Maybe<Scalars['String']['output']>;
  /** Target attribute for the menu item link. */
  target?: Maybe<Scalars['String']['output']>;
  /** Title attribute for the menu item link */
  title?: Maybe<Scalars['String']['output']>;
  /** The uri of the resource the menu item links to */
  uri?: Maybe<Scalars['String']['output']>;
  /** URL or destination of the menu item. */
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type MenuItemToMenuConnectionEdge = Edge & MenuConnectionEdge & OneToOneConnection & {
  __typename?: 'MenuItemToMenuConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: Menu;
};
// ...
export type NickelBronxHeader = AcfOptionsPage & Node & WithAcfConfig & {
  __typename?: 'NickelBronxHeader';
  /** Fields of the Config ACF Field Group */
  config?: Maybe<Config>;
  /** The globally unique ID for the object */
  id: Scalars['ID']['output'];
  menuTitle?: Maybe<Scalars['String']['output']>;
  pageTitle?: Maybe<Scalars['String']['output']>;
  parentId?: Maybe<Scalars['String']['output']>;
};
// ...
export type RootQueryToMenuItemConnection = Connection & MenuItemConnection & {
  __typename?: 'RootQueryToMenuItemConnection';
  /** Edges for the RootQueryToMenuItemConnection connection */
  edges: Array<RootQueryToMenuItemConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<MenuItem>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToMenuItemConnectionPageInfo;
};
// ...
export type GetHomePageQuery = { __typename?: 'RootQuery', generalSettings?: { __typename?: 'GeneralSettings', title?: string | null, description?: string | null } | null, primaryMenuItems?: { __typename?: 'RootQueryToMenuItemConnection', nodes: Array<{ __typename?: 'MenuItem', id: string, uri?: string | null, path?: string | null, label?: string | null, parentId?: string | null, cssClasses?: Array<string | null> | null, menu?: { __typename?: 'MenuItemToMenuConnectionEdge', node: { __typename?: 'Menu', name?: string | null } } | null }> } | null, nickelBronxHeader?: { __typename?: 'NickelBronxHeader', config?: { __typename?: 'Config', ctaText?: string | null, ctaLink?: string | null, brandLogo?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', sourceUrl?: string | null } } | null } | null } | null, page?: { __typename?: 'Page', editorBlocks?: Array<{ __typename?: 'AcfHeroSection', clientId?: string | null, heroSectionBlocks?: { __typename?: 'HeroSectionBlocks', heroDescription?: string | null, heroTitle?: string | null, heroCta?: { __typename?: 'HeroSectionBlocksHeroCta', ctaText?: string | null, ctaLink?: { __typename?: 'AcfLink', target?: string | null, title?: string | null, url?: string | null } | null } | null } | null } | { __typename?: 'AcfHeroVideoSection' } | { __typename?: 'AcfWhoWeAreSection' } | { __typename?: 'CoreArchives' } | { __typename?: 'CoreAudio' } | { __typename?: 'CoreAvatar' } | { __typename?: 'CoreBlock' } | { __typename?: 'CoreButton' } | { __typename?: 'CoreButtons' } | { __typename?: 'CoreCalendar' } | { __typename?: 'CoreCategories' } | { __typename?: 'CoreCode' } | { __typename?: 'CoreColumn' } | { __typename?: 'CoreColumns' } | { __typename?: 'CoreCommentAuthorName' } | { __typename?: 'CoreCommentContent' } | { __typename?: 'CoreCommentDate' } | { __typename?: 'CoreCommentEditLink' } | { __typename?: 'CoreCommentReplyLink' } | { __typename?: 'CoreCommentTemplate' } | { __typename?: 'CoreComments' } | { __typename?: 'CoreCommentsPagination' } | { __typename?: 'CoreCommentsPaginationNext' } | { __typename?: 'CoreCommentsPaginationNumbers' } | { __typename?: 'CoreCommentsPaginationPrevious' } | { __typename?: 'CoreCommentsTitle' } | { __typename?: 'CoreCover' } | { __typename?: 'CoreDetails' } | { __typename?: 'CoreEmbed' } | { __typename?: 'CoreFile' } | { __typename?: 'CoreFootnotes' } | { __typename?: 'CoreFreeform' } | { __typename?: 'CoreGallery' } | { __typename?: 'CoreGroup' } | { __typename?: 'CoreHeading' } | { __typename?: 'CoreHomeLink' } | { __typename?: 'CoreHtml' } | { __typename?: 'CoreImage' } | { __typename?: 'CoreLatestComments' } | { __typename?: 'CoreLatestPosts' } | { __typename?: 'CoreLegacyWidget' } | { __typename?: 'CoreList' } | { __typename?: 'CoreListItem' } | { __typename?: 'CoreLoginout' } | { __typename?: 'CoreMediaText' } | { __typename?: 'CoreMissing' } | { __typename?: 'CoreMore' } | { __typename?: 'CoreNavigation' } | { __typename?: 'CoreNavigationLink' } | { __typename?: 'CoreNavigationSubmenu' } | { __typename?: 'CoreNextpage' } | { __typename?: 'CorePageList' } | { __typename?: 'CorePageListItem' } | { __typename?: 'CoreParagraph' } | { __typename?: 'CorePattern' } | { __typename?: 'CorePostAuthor' } | { __typename?: 'CorePostAuthorBiography' } | { __typename?: 'CorePostAuthorName' } | { __typename?: 'CorePostComments' } | { __typename?: 'CorePostCommentsForm' } | { __typename?: 'CorePostContent' } | { __typename?: 'CorePostDate' } | { __typename?: 'CorePostExcerpt' } | { __typename?: 'CorePostFeaturedImage' } | { __typename?: 'CorePostNavigationLink' } | { __typename?: 'CorePostTemplate' } | { __typename?: 'CorePostTerms' } | { __typename?: 'CorePostTitle' } | { __typename?: 'CorePreformatted' } | { __typename?: 'CorePullquote' } | { __typename?: 'CoreQuery' } | { __typename?: 'CoreQueryNoResults' } | { __typename?: 'CoreQueryPagination' } | { __typename?: 'CoreQueryPaginationNext' } | { __typename?: 'CoreQueryPaginationNumbers' } | { __typename?: 'CoreQueryPaginationPrevious' } | { __typename?: 'CoreQueryTitle' } | { __typename?: 'CoreQuote' } | { __typename?: 'CoreReadMore' } | { __typename?: 'CoreRss' } | { __typename?: 'CoreSearch' } | { __typename?: 'CoreSeparator' } | { __typename?: 'CoreShortcode' } | { __typename?: 'CoreSiteLogo' } | { __typename?: 'CoreSiteTagline' } | { __typename?: 'CoreSiteTitle' } | { __typename?: 'CoreSocialLink' } | { __typename?: 'CoreSocialLinks' } | { __typename?: 'CoreSpacer' } | { __typename?: 'CoreTable' } | { __typename?: 'CoreTagCloud' } | { __typename?: 'CoreTemplatePart' } | { __typename?: 'CoreTermDescription' } | { __typename?: 'CoreTextColumns' } | { __typename?: 'CoreVerse' } | { __typename?: 'CoreVideo' } | { __typename?: 'CoreWidgetGroup' } | { __typename?: 'SafeSvgSvgIcon' } | null> | null } | null };